[ #22056 ] Port internal time handling from QDateTime to time_t
46% of the load time was spent converting time to string to time_t ot QDateTime or vice versa. 95% of this can be avoided quite easily, so I will. While not very nice from a C++ point of view, a 46% bottleneck is just too bad to ignore.

+++ src/sflphoneaccessibility.cpp
@@ -94,7 +94,7 @@ void SFLPhoneAccessibility::currentCallDetails()
          else if (!call->getCallNumber().isEmpty())
             toSay += i18n(", the phone number is %1 ",numberToDigit(call->getCallNumber()));
 
-         int nSec = QDateTime::fromTime_t(call->getStartTimeStamp().toInt()).time().secsTo( QTime::currentTime() );
+         int nSec = QDateTime::fromTime_t(call->getStartTimeStamp()).time().secsTo( QTime::currentTime() );
          if (nSec>0)
             toSay += i18n(" and you have been talking since %1 seconds",nSec );
